Fix breaks from initial checkin - https://trac.osgeo.org/gdal/ticket/6191
git-svn-id: https://svn.osgeo.org/gdal/trunk@32096 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/db2/ogrdb2datasource.cpp
@@ -130,7 +130,7 @@ OGRLayer *OGRDB2DataSource::GetLayerByName( const char* pszLayerName )
     const char* pszDotPos = strstr(pszLayerNameUpper,".");
     if ( pszDotPos != NULL )
     {
-        int length = pszDotPos - pszLayerNameUpper;
+        int length = static_cast<int>(pszDotPos - pszLayerNameUpper);
         pszSchemaName = (char*)CPLMalloc(length+1);
         strncpy(pszSchemaName, pszLayerNameUpper, length);
         pszSchemaName[length] = '\0';
@@ -262,7 +262,7 @@ OGRLayer * OGRDB2DataSource::ICreateLayer( const char * pszLayerName,
     const char* pszDotPos = strstr(pszLayerName,".");
     if ( pszDotPos != NULL )
     {
-        int length = pszDotPos - pszLayerName;
+        int length = static_cast<int>(pszDotPos - pszLayerName);
         pszSchemaName = (char*)CPLMalloc(length+1);
         strncpy(pszSchemaName, pszLayerName, length);
         pszSchemaName[length] = '\0';
@@ -492,7 +492,7 @@ int OGRDB2DataSource::ParseValue(char** pszValue, char* pszSource,
                                  const char* pszKey, int nStart, int nNext,
                                  int nTerm, int bRemove)
 {
-    int nLen = strlen(pszKey);
+    int nLen = static_cast<int>(strlen(pszKey));
     if ((*pszValue) == NULL && nStart + nLen < nNext &&
             EQUALN(pszSource + nStart, pszKey, nLen))
     {
@@ -575,7 +575,7 @@ int OGRDB2DataSource::Open( const char * pszNewName, int bUpdate,
     char* pszConnectionName = CPLStrdup(pszNewName + strlen(DB2ODBC_PREFIX));
     char* pszDriver = NULL;
     int nCurrent, nNext, nTerm;
-    nCurrent = nNext = nTerm = strlen(pszConnectionName);
+    nCurrent = nNext = nTerm = static_cast<int>(strlen(pszConnectionName));
 
     while (nCurrent > 0)
     {
@@ -652,7 +652,7 @@ int OGRDB2DataSource::Open( const char * pszNewName, int bUpdate,
                 {
                     *pos = '\0';
                     pszGeomColumnName = pos+1;
-                    int len = strlen(pszGeomColumnName);
+                    int len = static_cast<int>(strlen(pszGeomColumnName));
                     if (len > 0)
                         pszGeomColumnName[len - 1] = '\0';
                 }
